pyproject.toml
[tool.poetry]
name = "robocar-oak-camera"
version = "0.0.0"
description = "Mqtt gateway for oak-lite device"
authors = ["Cyrille Nofficial <cynoffic@cyrilix.fr>"]
readme = "README.md"
packages = [
    { include = "camera" },
]

[tool.poetry.dependencies]
python = "^3.10"
paho-mqtt = "^1.6.1"
depthai = "^2.17.4.0"
protobuf3 = "^0.2.1"
google = "^3.0.0"
blobconverter = "^1.3.0"
protobuf = "^4.21.8"
opencv-python-headless = "^4.6.0.66"
robocar-protobuf = { version = "^1.1.1", source = "robocar" }


[tool.poetry.group.test.dependencies]
pytest = "^7.1.3"
pytest-mock = "^3.10.0"


[tool.poetry.group.dev.dependencies]
pylint = "^2.15.4"
mypy = "^0.982"
types-paho-mqtt = "^1.6.0.1"
types-protobuf = "^3.20.4.2"


[[tool.poetry.source]]
name = "robocar"
url = "https://git.cyrilix.bzh/api/packages/robocars/pypi/simple"
default = false
secondary = false

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry.scripts]
rc-oak-camera = 'camera.cli:execute_from_command_line'

[tool.poetry-dynamic-versioning]
enable = true
style = 'semver'
vcs = 'git'
dirty = true
bump = true
